@@ -12,10 +12,10 @@
   "private": true,
   "dependencies": {
     "@angular/animations": "^14.0.0",
-    "@angular/cdk": "^13.3.9",
+    "@angular/cdk": "^14.0.0",
     "@angular/common": "^14.0.0",
     "@angular/compiler": "^14.0.0",
-    "@angular/core": "^13.3.11",
+    "@angular/core": "^14.0.0",
     "@angular/forms": "^14.0.0",
     "@angular/material": "^13.3.9",
     "@angular/platform-browser": "^14.0.0",
@@ -27,8 +27,8 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.3.7",
-    "@angular/cli": "^13.3.7",
+    "@angular-devkit/build-angular": "^14.0.0",
+    "@angular/cli": "^14.0.0",
     "@angular/compiler-cli": "^14.0.0",
     "@angular/language-service": "^14.0.0",
     "@types/jasmine": "^3.10.6",
